separate nav2 library (#127)
