+++ examples/example_scenario_control/scenarios/nav2_pose_2.osc
@@ -1,4 +1,5 @@
 import osc.ros
+import osc.nav2
 
 scenario nav2_simulation_nav_to_pose_2:
     turtlebot4: differential_drive_robot
