+++ libs/scenario_execution_floorplan_dsl/example/example.osc
@@ -1,6 +1,7 @@
 import osc.helpers
 import osc.floorplan_dsl
 import osc.ros
+import osc.nav2
 
 scenario test:
     timeout(120s)
