+++ libs/scenario_execution_nav2/MANIFEST.in
@@ -1,0 +1,1 @@
+include scenario_execution_nav2/lib_osc/*.osc
