+++ libs/scenario_execution_nav2/README.md
@@ -1,0 +1,8 @@
+# Scenario Execution Library for Nav2
+
+The `scenario_execution_nav2` package provides actions to interact with the [nav2](https://docs.nav2.org/) navigation stack.
+
+It provides the following scenario execution library:
+
+- `nav2.osc`: Actions specific to nav2 navigation stack
+
