+++ libs/scenario_execution_nav2/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/scenario_execution_nav2
+[install]
+install_scripts=$base/lib/scenario_execution_nav2
