+++ scenario_coverage/scenarios/test_fault_injection_noise.osc
@@ -1,5 +1,6 @@
 import osc.helpers
 import osc.ros
+import osc.nav2
 
 scenario nav2_simulation_fault_injection:
     timeout(240s)
