+++ scenario_coverage/scenarios/test_nav_to_pose.osc
@@ -1,5 +1,6 @@
 import osc.helpers
 import osc.ros
+import osc.nav2
 
 scenario nav2_simulation_nav_to_pose:
     timeout(240s)
