+++ scenario_execution_ros/package.xml
@@ -17,7 +17,6 @@
   <exec_depend>py_trees_ros</exec_depend>
   <exec_depend>py_trees_ros_interfaces</exec_depend>
   <exec_depend>visualization_msgs</exec_depend>
-  <exec_depend>nav2_simple_commander</exec_depend>
   <exec_depend>xacro</exec_depend>
   <exec_depend>rcl_interfaces</exec_depend>
   <exec_depend>scenario_status</exec_depend>
