+++ scenario_execution_ros/scenario_execution_ros/actions/tf_close_to.py
@@ -26,7 +26,7 @@ from geometry_msgs.msg import PoseStamped
 from tf2_ros.buffer import Buffer
 from tf2_ros import TransformException  # pylint: disable= no-name-in-module
 
-from scenario_execution_ros.actions.nav2_common import NamespacedTransformListener
+from .common import NamespacedTransformListener
 from scenario_execution.actions.base_action import BaseAction
